Add /publicapi/contact-us-config/ endpoint
