+++ app/client/components/contactus/contactus.tsx
@@ -4,9 +4,9 @@ import { palette, space } from "@guardian/src-foundations";
 import { headline, textSans } from "@guardian/src-foundations/typography";
 import { RouteComponentProps } from "@reach/router";
 import React, { useEffect, useState } from "react";
+import { contactUsConfig } from "../../../server/contactUsConfig";
 import { minWidth } from "../../styles/breakpoints";
 import { trackEvent } from "../analytics";
-import { contactUsConfig } from "./contactUsConfig";
 import { ContactUsForm } from "./contactUsForm";
 import { ContactUsHeader } from "./contactUsHeader";
 import { ContactUsPageContainer } from "./contactUsPageContainer";
