+++ app/server/contactUsConfig.ts
@@ -1,27 +1,4 @@
-interface BaseTopic {
-  id: string;
-  name: string;
-  selfServiceBox?: SelfServiceBox;
-  editableSubjectLine?: boolean;
-}
-
-interface SelfServiceBox {
-  text: string;
-  linkText: string;
-  href: string;
-  noForm?: boolean;
-}
-
-export interface SubTopic extends BaseTopic {
-  subsubtopics?: BaseTopic[];
-  subsubTopicsTitle?: string;
-}
-
-export interface Topic extends BaseTopic {
-  enquiryLabel: string;
-  subtopics?: SubTopic[];
-  subTopicsTitle?: string;
-}
+import { Topic } from "../shared/contactUsTypes";
 
 export const contactUsConfig: Topic[] = [
   {
