+++ app/server/routes/index.ts
@@ -4,3 +4,4 @@ export { default as profile } from "./profile";
 export { default as frontend } from "./frontend";
 export { default as productsProvider } from "./products";
 export { default as idapi } from "./idapi";
+export { default as publicApi } from "./publicApi";
